requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
requests==2.26.0
python-dotenv==0.19.0
schedule==1.1.0
pytz==2023.3 